+++ oci/security.tf
@@ -4,8 +4,8 @@
 # Name.......: security.tf
 # Author.....: Stefan Oehrli (oes) stefan.oehrli@oradba.ch
 # Editor.....: Stefan Oehrli
-# Date.......: 2024.10.17
-# Revision...: 1.0.0
+# Date.......: 2024.10.23
+# Revision...: 0.3.2
 # Purpose....: Define security lists for Virtual Cloud Network (VCN) resources 
 #              in the LAB environment, including rules for public and private 
 #              subnets to allow or restrict traffic.
@@ -24,11 +24,10 @@
 # This resource defines the security list for public subnets in each lab 
 # environment. It allows outbound HTTP and HTTPS traffic and can be customized 
 # with additional rules as needed.
-resource "oci_core_security_list" "public_security_list" {
-  count = var.numberOf_labs # Create one security list per lab
-
-  compartment_id = oci_identity_compartment.lab-compartment[count.index].id # Compartment for each lab
-  vcn_id         = oci_core_vcn.vcn[count.index].id                         # Associate security list with the VCN
+resource "oci_core_default_security_list" "public_security_list" {
+  count                      = var.numberOf_labs # Create one security list per lab
+  manage_default_resource_id = oci_core_vcn.vcn[count.index].default_security_list_id
+  compartment_id             = oci_identity_compartment.lab-compartment[count.index].id # Compartment for each lab
 
   # Define display name using lab-specific variables such as region, environment, and resource prefix
   display_name = format("sl-pub-%s-%s-%s-%02d", lower(local.current_region_key), lower(var.environment_code), lower(local.resource_prefix_shortname), count.index)
@@ -36,6 +35,11 @@ resource "oci_core_security_list" "public_security_list" {
   # ------------------------------------------------------------------------------
   # Egress Rules (Outbound Traffic)
   # ------------------------------------------------------------------------------
+  egress_security_rules {
+    protocol    = "all"              # Allow all protocols
+    destination = var.vcn_cidr_block # Allow all traffic within the VCN's CIDR block
+    description = "Allow all outbound traffic within the subnet"
+  }
   egress_security_rules {
     protocol    = "6"                           # TCP protocol
     destination = "0.0.0.0/0"                   # Allow outbound traffic to all destinations