+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - change password length to 18 characters
 - use default dhcp option for public network
 - add comments to logging resource configuration
+- use default security list for public network
 
 ### Fixed